Refactor connectivity package: extract HTTP client and cookie handling logic
