+++ internal/collector/connectivity/fetch.go
